Isolated Processes 2: Redirect thread_info I/O and ptrace
thread_info is complex enough to redirect that it could use its own
PR for review.
