+++ src/linux/mod.rs
@@ -1,19 +1,24 @@
 // `WriterError` is large and clippy doesn't like that, but not a huge deal atm
 #![allow(clippy::result_large_err)]
 
-#[cfg(target_os = "android")]
-mod android;
+pub use maps_reader::LINUX_GATE_LIBRARY_NAME;
+
 pub mod app_memory;
-pub(crate) mod auxv;
 pub mod crash_context;
-mod dso_debug;
-mod dumper_cpu_info;
 pub mod maps_reader;
 pub mod minidump_writer;
 pub mod module_reader;
 pub mod process_reader;
-mod serializers;
 pub mod thread_info;
 
-pub use maps_reader::LINUX_GATE_LIBRARY_NAME;
+pub(crate) mod auxv;
+
+mod dso_debug;
+mod dumper_cpu_info;
+mod process_inspection;
+mod serializers;
+
+#[cfg(target_os = "android")]
+mod android;
+
 pub type Pid = i32;
