+++ src/linux/thread_info/arm.rs
@@ -1,24 +1,7 @@
 use {
-    super::{Pid, PtraceRequestType, ThreadInfoError},
+    super::{Pid, ProcessInspector, ThreadInfoError, regs::*},
     crate::minidump_cpu::RawContextCPU,
 };
-
-type Result<T> = std::result::Result<T, ThreadInfoError>;
-
-// Not defined by libc because this works only for cores support VFP
-#[allow(non_camel_case_types)]
-#[repr(C)]
-#[derive(Debug, Eq, Hash, PartialEq, Copy, Clone, Default)]
-pub struct user_fpregs_struct {
-    pub fpregs: [u64; 32],
-    pub fpscr: u32,
-}
-
-#[repr(C)]
-#[derive(Debug, Eq, Hash, PartialEq, Copy, Clone, Default)]
-pub struct user_regs_struct {
-    uregs: [u32; 18],
-}
 
 #[derive(Debug)]
 pub struct ThreadInfoArm {
@@ -30,16 +13,6 @@ pub struct ThreadInfoArm {
 }
 
 impl ThreadInfoArm {
-    fn getfpregs(pid: Pid) -> Result<user_fpregs_struct> {
-        const NT_ARM_VFP: usize = 0x400;
-        super::ptrace_getregset(NT_ARM_VFP, pid)
-    }
-
-    fn getregs(pid: Pid) -> Result<user_regs_struct> {
-        const PTRACE_GETREGS: PtraceRequestType = 12;
-        unsafe { super::ptrace_getregs::<user_regs_struct>(PTRACE_GETREGS, pid) }
-    }
-
     pub fn get_instruction_pointer(&self) -> usize {
         self.regs.uregs[15] as usize
     }
@@ -54,10 +27,18 @@ impl ThreadInfoArm {
         out.float_save.regs = self.fpregs.fpregs;
     }
 
-    pub fn create(_pid: Pid, tid: Pid) -> Result<Self> {
-        let (ppid, tgid) = super::get_ppid_and_tgid(tid)?;
-        let regs = Self::getregs(tid)?;
-        let fpregs = Self::getfpregs(tid).unwrap_or(Default::default());
+    pub fn create(
+        process_inspector: &ProcessInspector,
+        _pid: Pid,
+        tid: Pid,
+    ) -> Result<Self, ThreadInfoError> {
+        let (ppid, tgid) = super::get_ppid_and_tgid(process_inspector, tid)?;
+        let regs = process_inspector
+            .get_gen_regs(tid)
+            .map_err(ThreadInfoError::PtraceError)?;
+        let fpregs = process_inspector
+            .get_fp_regs(tid)
+            .map_err(ThreadInfoError::PtraceError)?;
 
         let stack_pointer = regs.uregs[13] as usize;
